requirements.txt
# Core (Django 5.x supports Python 3.11; Django 6.x requires Python 3.12+)
Django>=5.1,<6.0

# Multi-tenancy
django-tenants>=3.7.0,<4.0

# REST API
djangorestframework>=3.15.0,<4.0

# CORS
django-cors-headers>=4.6.0,<5.0

# Task Queue
celery>=5.4.0,<6.0

# PostgreSQL adapter (required by django-tenants)
psycopg2-binary>=2.9.9,<3.0
